moved component to different file
